Extract recursion functions for reuse
