+++ backend/src/apis/CMT/evaluate/evaluate.py
@@ -5,6 +5,7 @@ from keras import preprocessing as pp
 from concurrent.futures import ThreadPoolExecutor
 
 from apis.utils.directory import dire
+from backend.src.apis.utils.recursive import recursion
 
 
 def evaluate(model, labels):
@@ -43,17 +44,6 @@ def evaluate(model, labels):
             #                 results[key][lot_plate][last_fol] = len(files)
 
     return results
-
-
-def recursion(folder_dict, path_list, function):
-    head, *tail = path_list
-    if 1 < len(path_list):
-        if head not in folder_dict:
-            folder_dict[head] = recursion({}, tail, function)
-        else:
-            folder_dict[head].update(recursion(folder_dict[head], tail, function))
-    else:
-        return {head: function}
 
 
 def predict(root, model, labels):
